+++ derive/src/enum_derive/args.rs
@@ -1,8 +1,12 @@
 use proc_macro2::{token_stream::IntoIter, Ident, TokenStream, TokenTree};
+use quote::ToTokens;
 use smallvec::smallvec;
-use syn::Path;
+use syn::{Path, Result};
 
-use crate::utils::*;
+use crate::utils::Stack;
+
+// =============================================================================
+// Arg
 
 #[derive(Clone)]
 pub(super) enum Arg {
@@ -40,6 +44,21 @@ impl From<Path> for Arg {
     }
 }
 
+// =============================================================================
+// Parse
+
+macro_rules! arg_err {
+    ($msg:expr) => {
+        syn::Error::new($msg.span(), format!("invalid arguments: {}", $msg))
+    };
+    ($span:expr, $msg:expr) => {
+        syn::Error::new($span.span(), format!("invalid arguments: {}", $msg))
+    };
+    ($span:expr, $($tt:tt)*) => {
+        arg_err!($span, format!($($tt)*))
+    };
+}
+
 pub(super) fn parse_args(args: TokenStream) -> Result<Stack<(String, Arg)>> {
     fn push(args: &mut Stack<(String, Arg)>, arg: Arg) {
         args.push((arg.to_trimed_string(), arg))
@@ -55,9 +74,9 @@ pub(super) fn parse_args(args: TokenStream) -> Result<Stack<(String, Arg)>> {
             TokenTree::Punct(p) => match p.as_char() {
                 ',' => {}
                 ':' => push(&mut args, parse_path(smallvec![p.into()], &mut iter)?),
-                _ => Err(invalid_args!("{}`{}`", ERR, p))?,
+                _ => Err(arg_err!(p, "{}`{}`", ERR, p))?,
             },
-            _ => Err(invalid_args!("{}`{}`", ERR, tt))?,
+            _ => Err(arg_err!(tt, "{}`{}`", ERR, tt))?,
         }
     }
 
@@ -73,7 +92,7 @@ fn parse_path(mut path: Stack<TokenTree>, iter: &mut IntoIter) -> Result<Arg> {
     }
 
     syn::parse2(path.into_iter().collect())
-        .map_err(|e| invalid_args!(e))
+        .map_err(|e| arg_err!(e))
         .map(Arg::Path)
 }
 
@@ -85,8 +104,8 @@ fn path_or_ident(ident: Ident, tt: Option<TokenTree>, iter: &mut IntoIter) -> Re
         Some(TokenTree::Punct(p)) => match p.as_char() {
             ',' => Ok(ident.into()),
             ':' => parse_path(smallvec![ident.into(), p.into()], iter),
-            _ => Err(invalid_args!("{}`{}`", ERR, p)),
+            _ => Err(arg_err!(p, "{}`{}`", ERR, p)),
         },
-        Some(tt) => Err(invalid_args!("{}`{}`", ERR, tt)),
+        Some(tt) => Err(arg_err!(tt, "{}`{}`", ERR, tt)),
     }
 }